pyproject.toml
[project]
name = "fedpoffa-cbt-backend"
version = "0.1.0"
description = "FEDPOFFA Computer-Based Testing System Backend"
authors = [
    {name = "FEDPOFFA IT Team"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "email-validator (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.0"
pre-commit = "^4.2.0"

